chore: setup docker compose for fossil api
Setup docker and docker compose for fossil api, allowing it to be ran
locally with a local db instance running in docker.

Also added auto migration with this docker setup.

We can later re-use this for deployment if needed, but we should remove
the db component and utilize an external db.