feat: separate the indexer & offchain db (#81)
* feat: separate the indexer & offchain db

For the sake of data separation for the indexer & the data used for
offchain processor (this repo), add wrappers to constraint indexer
accessors to use only indexer db and offchain operations to use the
offchain db.

* feat: modify db crate to use offchain db

Since we used to have one db for everything, the sqlx prepare & macro
usage is convenient. However since we will be having 2 dbs that confuses
the sqlx macro and there is currently no easy way to specify to each
macro to listen to different db url via env vars. As such this removes
macro and sqlx prepared usage for the indexer db, and only uses those
for the offchain db.

If we think that we would want to have a unified api used for both db
accesses then we can also switch to all non-macro usage.